Code Cleaning

+++ tabugroup/Tabu.hpp
@@ -1,11 +1,8 @@
 #ifndef Tabu_hpp
 #define Tabu_hpp
 
-#include <stdio.h>
-#include "../graphw.hpp"
 #include "../Solution.hpp"
 #include "../TimeController.hpp"
-#include <ctime>
 
 class Tabu{
 public:
@@ -23,7 +20,7 @@ private:
     /*FUNZIONI**/
     
     //Cerca la lista Tabu per vedere se una mossa è proibita
-    bool isTabu(long int id, int color);
+    bool isTabu(int id, int color);
     bool isTabu(G::Graph& g,G::Vertex v, int color);
     
     //Inserire mossa nella tabu list
@@ -32,14 +29,14 @@ private:
     
     //
     void tabuSimpleKempeWrapper(G::Graph& g ,G::Vertex v, int color);
-    void tabuSimpleKempe(G::Graph& g ,G::Vertex v, int color,std::unordered_set<long int>& visitedNodes);
+    void tabuSimpleKempe(G::Graph& g ,G::Vertex v, int color,std::unordered_set<int>& visitedNodes);
     void resetTabuList();
     
     //Si segna le mosse già esplorate in una matrice apposita
     int tabuKempeMovePenaltyWrapper(G::Graph& g, G::Vertex v, int color);
-    int tabuKempeMovePenalty(G::Graph& g, G::Vertex v, int color,  std::unordered_set<long int>& visitedNodes);
+    int tabuKempeMovePenalty(G::Graph& g, G::Vertex v, int color,  std::unordered_set<int>& visitedNodes);
     void updatePenaltyWrapper(G::Graph& g, G::Vertex v, int color, int penalty);
-    void updatePenalty(G::Graph& g, G::Vertex v, int color, int penalty,std::unordered_set<long int>& visitedNodes);
+    void updatePenalty(G::Graph& g, G::Vertex v, int color, int penalty,std::unordered_set<int>& visitedNodes);
     void updateMoveMatrix(G::Graph& g); //Aggiorna la matrice delle mosse
     void renderFalseMoves(G::Graph& g,G::Vertex v); //Renders all moves associated with
     //node v false
@@ -48,7 +45,7 @@ private:
     
     //Aggiorna le penalita di un nodo
     void updateNodePenalties(G::Graph& g,G::Vertex v);
-    std::pair<long int,int> bestMove();
+    std::pair<int,int> bestMove();
     
     //Perturba ma anche si ricorda delle mosse precalcolate
     void tabuPerturbate(G::Graph& g, int q,int eta, int tmax);
@@ -57,10 +54,10 @@ private:
     //La scelta è fatta in maniera tale che questi due nodi appartengono a due catene
     //Kempe diverse, per cui si effettua un doppio kempe
     //tuple ritorna nodo1,nodo2, colore
-    std::tuple<long int,long int,int> bestDoubleKempe(G::Graph& g);
+    std::tuple<int,int,int> bestDoubleKempe(G::Graph& g);
     
    //Accetta due nodi e il colore al quale devono venir swappati
-    int doubleKempeSwap(G::Graph&,long int,long int,int);
+    int doubleKempeSwap(G::Graph&,int,int,int);
     
     //Swappa due colori randomici
     void swapColors(G::Graph& g);
@@ -76,12 +73,8 @@ private:
     int teta; //Lunghezza della tabu list
     //Lista delle mosse proibite, il primo intero è l'id del vettore
     //Il secondo elemento è il colore al quale non si può muovere
-    std::deque<std::pair<long int,int>> tabuList;
+    std::deque<std::pair<int,int>> tabuList;
     int studentNum;
-    
-
-
-
     
 };
